Issues 804 : Update projectList
